Convert the Dependents query to a new Fencers query
