+++ src/components/FencersManager/components/EditFencerDialog/index.ts
@@ -1,0 +1,1 @@
+export * from "./EditFencerDialog"
