+++ src/components/index.ts
@@ -13,3 +13,4 @@ export * from "./Forms"
 export * from "./PistePanel"
 export * from "./PaymentMethod"
 export * from "./AuthorizedApolloProvider"
+export * from "./FencersManager"
